fix(接口测试): 修复json-schema转json类型报错的缺陷
--bug=1027618 --user=王孝刚
【接口测试】github#25412，接口测试json请求体使用json-schema格式化之后再关闭schema，请求体内容发生了很奇怪的变化
https://www.tapd.cn/55049933/s/1390774

+++ api-test/backend/src/main/java/io/metersphere/api/exec/generator/JSONSchemaParser.java
@@ -125,7 +125,7 @@ public class JSONSchemaParser {
                 if (FormatterUtil.isNumber(FormatterUtil.getStrValue(object))) {
                     String value = FormatterUtil.getElementValue(object).getAsString();
                     if (value.indexOf(".") == -1) {
-                        concept.put(propertyName, Integer.valueOf(value));
+                        concept.put(propertyName, Long.valueOf(value));
                     } else {
                         concept.put(propertyName, new BigDecimal(value));
                     }
